refactor(apps/kampus): split baseUrl and product url folder
